refactor: 使用 floating-ui 重构 tooltip
